Add missing cases in fuse
When fusing two trees, their subtrees may be just leaves,
resulting in the recursive call to fuse returning a leaf.
This case does not introduce any new logic, but has to be
handled in the patterns.

minimal example where the missing case results in an error:

delete 3 (foldr insert empty [1..4])